+++ packages/components/src/components.d.ts
@@ -1708,9 +1708,9 @@ export namespace Components {
          */
         "color": Props.BalStageColor;
         /**
-          * class to set the container width
+          * Defines content width of the stage
          */
-        "containerClass": string;
+        "containerSize": '' | 'is-fluid' | 'is-detail-page' | 'is-compact' | 'is-blog-page' | 'is-wide';
         /**
           * If true the Baloise Shape is set
          */
@@ -1722,7 +1722,15 @@ export namespace Components {
         /**
           * sets text color to white for images and dark backgrounds (optional)
          */
-        "inverted": undefined | boolean;
+        "inverted"?: boolean;
+        /**
+          * Shape Rotation
+         */
+        "shapeRotation"?: Props.BalShapeRotation;
+        /**
+          * Shape Variation
+         */
+        "shapeVariation"?: Props.BalShapeVariation;
         /**
           * Defines size of the stage
          */
@@ -1735,10 +1743,6 @@ export namespace Components {
         "href": string;
     }
     interface BalStageBody {
-        /**
-          * class to set the container width
-         */
-        "containerClass": string;
     }
     interface BalStageFoot {
     }
@@ -4693,9 +4697,9 @@ declare namespace LocalJSX {
          */
         "color"?: Props.BalStageColor;
         /**
-          * class to set the container width
+          * Defines content width of the stage
          */
-        "containerClass"?: string;
+        "containerSize"?: '' | 'is-fluid' | 'is-detail-page' | 'is-compact' | 'is-blog-page' | 'is-wide';
         /**
           * If true the Baloise Shape is set
          */
@@ -4707,7 +4711,15 @@ declare namespace LocalJSX {
         /**
           * sets text color to white for images and dark backgrounds (optional)
          */
-        "inverted"?: undefined | boolean;
+        "inverted"?: boolean;
+        /**
+          * Shape Rotation
+         */
+        "shapeRotation"?: Props.BalShapeRotation;
+        /**
+          * Shape Variation
+         */
+        "shapeVariation"?: Props.BalShapeVariation;
         /**
           * Defines size of the stage
          */
@@ -4720,10 +4732,6 @@ declare namespace LocalJSX {
         "href": string;
     }
     interface BalStageBody {
-        /**
-          * class to set the container width
-         */
-        "containerClass"?: string;
     }
     interface BalStageFoot {
     }